requirements.txt
appscript==1.4.0
blinker==1.9.0
click==8.3.1
docx2pdf==0.1.8
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
python-docx==1.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.44
tqdm==4.67.1
typing_extensions==4.15.0
Werkzeug==3.1.4
